Add block permission billing delete

+++ server/controller/Api/block/flat.php
@@ -8,6 +8,7 @@ use Selpol\Controller\Api\Api;
 use Selpol\Entity\Model\Block\FlatBlock;
 use Selpol\Feature\Block\BlockFeature;
 use Selpol\Framework\Http\Response;
+use Selpol\Service\AuthService;
 use Selpol\Task\Tasks\Inbox\InboxFlatTask;
 use Selpol\Task\Tasks\Intercom\Flat\IntercomSyncFlatTask;
 
@@ -74,6 +75,12 @@ readonly class flat extends Api
     public static function DELETE(array $params): array|Response|ResponseInterface
     {
         $flatBlock = FlatBlock::findById($params['_id'], setting: setting()->nonNullable());
+/*
+ *             'block-flat-billing-delete' => '[Блокировка-Квартира] Удалить блокировку биллинга',
+            'block-subscriber-billing-delete' => '[Блокировка-Абонент] Удалить блокировку биллинга',
+ */
+        if ($flatBlock->status == BlockFeature::STATUS_BILLING && !container(AuthService::class)->checkScope('block-flat-billing-delete'))
+            return self::error('Не удалось удалить блокировку квартиры', 400);
 
         if ($flatBlock->delete()) {
             if ($flatBlock->service == BlockFeature::SERVICE_INTERCOM || $flatBlock->service == BlockFeature::SUB_SERVICE_CMS)
